Creates AggregatePreferenceFragment from AggregatePreferenceActivity
